Should fix the import check error
Added ipywidgets to setup.py, put the functon in a new file ux.py to attempt solving the circular import
