+++ setup.py
@@ -20,7 +20,7 @@ setuptools.setup(
         'Documentation': 'https://statwrap.readthedocs.io/'
     },
     author='Alexander Clark',
-    install_requires=['pandas','numpy','scipy', 'IPython', 'matplotlib'],
+    install_requires=['pandas','numpy','scipy', 'IPython', 'matplotlib', 'ipywidgets'],
     author_email='',  # consider adding your email
     packages=setuptools.find_packages(),
     zip_safe=False,
